+++ tests/evergreen/test_api.py
@@ -563,14 +563,14 @@ class TestCreatePatchDiff:
         mock_run.return_value = mock_stdout
 
         result = mocked_api.patch_from_diff(
-            "path", "params", "base", "task", "project", "description", "variant"
+            "path", "params", "base", "task", "project", "description", "variant", "build_id"
         )
 
         command = mock_run.call_args[0][0]
 
         assert (
             command
-            == "evergreen patch-file --diff-file path --description 'description' --param params --base base --tasks task --variants variant --project project -y -f"
+            == "evergreen patch-file --diff-file path --description 'description' --param params --base base --tasks task --variants variant --project project -y -f --param reuse_compile_from=build_id"
         )
         assert (
             result.url
@@ -584,14 +584,22 @@ class TestCreatePatchDiff:
         mock_run.return_value = mock_stdout
 
         result = mocked_api.patch_from_diff(
-            "path", "params", "base", "task", "project", "description", "variant", "author"
+            "path",
+            "params",
+            "base",
+            "task",
+            "project",
+            "description",
+            "variant",
+            "build_id",
+            "author",
         )
 
         command = mock_run.call_args[0][0]
 
         assert (
             command
-            == "evergreen patch-file --diff-file path --description 'description' --param params --base base --tasks task --variants variant --project project -y -f --author author"
+            == "evergreen patch-file --diff-file path --description 'description' --param params --base base --tasks task --variants variant --project project -y -f --param reuse_compile_from=build_id --author author"
         )
         assert (
             result.url
@@ -608,6 +616,27 @@ class TestCreatePatchDiff:
             mocked_api.patch_from_diff(
                 "path", "params", "base", "task", "project", "description", "variant"
             )
+
+    @patch("evergreen.api.subprocess.run",)
+    def test_patch_from_patch_id(self, mock_run, mocked_api):
+        mock_stdout = MagicMock()
+        mock_stdout.stderr = b"[evergreen] 2023/04/13 15:05:24 [p=info]: Patch successfully created.\n[evergreen] 2023/04/13 15:05:24 [p=info]: \n         ID : 64387ca457e85ac95a3da12f\n    Created : 2023-04-13 22:05:24.463 +0000 UTC\n    Description : Test enable profiling.\n      Build : https://evergreen.mongodb.com/patch/64387ca457e85ac95a3da12f?redirect_spruce_users=true\n     Status : created\n\n\n"
+        mock_run.return_value = mock_stdout
+
+        result = mocked_api.patch_from_patch_id(
+            "build_id", "params", "base", "task", "project", "description", "variant"
+        )
+
+        command = mock_run.call_args[0][0]
+
+        assert (
+            command
+            == "evergreen patch-file --diff-patchId build_id --description 'description' --param params --tasks task --variants variant --project project -y -f --param reuse_compile_from=build_id"
+        )
+        assert (
+            result.url
+            == "https://evergreen.mongodb.com/patch/64387ca457e85ac95a3da12f?redirect_spruce_users=true"
+        )
 
 
 class TestTaskApi(object):